Replace unsigned long long with uint64_t

+++ lonestar/analytics/distributed/bfs/bfs_push_cuda.cu
@@ -13,7 +13,7 @@ bool enable_lb = true;
 #include "bfs_push_cuda.cuh"
 static const int __tb_BFS = TB_SIZE;
 static const int __tb_FirstItr_BFS = TB_SIZE;
-__global__ void InitializeGraph(CSRGraph graph, unsigned int __begin, unsigned int __end, const uint32_t  local_infinity, unsigned long long local_src_node, uint32_t * p_dist_current, uint32_t * p_dist_old)
+__global__ void InitializeGraph(CSRGraph graph, unsigned int __begin, unsigned int __end, const uint32_t  local_infinity, uint64_t local_src_node, uint32_t * p_dist_current, uint32_t * p_dist_old)
 {
   unsigned tid = TID_1D;
   unsigned nthreads = TOTAL_THREADS_1D;
@@ -729,7 +729,7 @@ __global__ void BFSSanityCheck(CSRGraph graph, unsigned int __begin, unsigned in
   DGMax.thread_exit<cub::BlockReduce<uint32_t, TB_SIZE> >(DGMax_ts);
   // FP: "16 -> 17;
 }
-void InitializeGraph_cuda(unsigned int  __begin, unsigned int  __end, const uint32_t & local_infinity, unsigned long long local_src_node, struct CUDA_Context*  ctx)
+void InitializeGraph_cuda(unsigned int  __begin, unsigned int  __end, const uint32_t & local_infinity, uint64_t local_src_node, struct CUDA_Context*  ctx)
 {
   t_work.init_thread_work(ctx->gg.nnodes);
   dim3 blocks;
@@ -745,19 +745,19 @@ void InitializeGraph_cuda(unsigned int  __begin, unsigned int  __end, const uint
   check_cuda_kernel;
   // FP: "6 -> 7;
 }
-void InitializeGraph_allNodes_cuda(const uint32_t & local_infinity, unsigned long long local_src_node, struct CUDA_Context*  ctx)
+void InitializeGraph_allNodes_cuda(const uint32_t & local_infinity, uint64_t local_src_node, struct CUDA_Context*  ctx)
 {
   // FP: "1 -> 2;
   InitializeGraph_cuda(0, ctx->gg.nnodes, local_infinity, local_src_node, ctx);
   // FP: "2 -> 3;
 }
-void InitializeGraph_masterNodes_cuda(const uint32_t & local_infinity, unsigned long long local_src_node, struct CUDA_Context*  ctx)
+void InitializeGraph_masterNodes_cuda(const uint32_t & local_infinity, uint64_t local_src_node, struct CUDA_Context*  ctx)
 {
   // FP: "1 -> 2;
   InitializeGraph_cuda(ctx->beginMaster, ctx->beginMaster + ctx->numOwned, local_infinity, local_src_node, ctx);
   // FP: "2 -> 3;
 }
-void InitializeGraph_nodesWithEdges_cuda(const uint32_t & local_infinity, unsigned long long local_src_node, struct CUDA_Context*  ctx)
+void InitializeGraph_nodesWithEdges_cuda(const uint32_t & local_infinity, uint64_t local_src_node, struct CUDA_Context*  ctx)
 {
   // FP: "1 -> 2;
   InitializeGraph_cuda(0, ctx->numNodesWithEdges, local_infinity, local_src_node, ctx);
